# LEED v5 Credit Automation V1/leed-platform/backend/api_spec.yaml
openapi: 3.0.3
info:
  title: LEED AI Platform API
  version: 1.0.0
  description: API for LEED v5 credit automation platform
servers:
  - url: https://api.leedai.io/v1
    description: Production
  - url: https://staging-api.leedai.io/v1
    description: Staging
security:
  - bearerAuth: []
paths:
...
  /projects:
    get:
      summary: List projects
      tags: [Projects]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: List of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectList'

    post:
      summary: Create new project
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        201:
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/{projectId}:
    get:
      summary: Get project details
      tags: [Projects]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetail'

    patch:
      summary: Update project
      tags: [Projects]
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        200:
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        building_type:
          type: string
        leed_version:
          type: string
        rating_system:
          type: string
        target_level:
          type: string
        location:
          $ref: '#/components/schemas/Location'
        created_at:
          type: string
          format: date-time
    ProjectDetail:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            credits_completed:
              type: integer
            credits_total:
              type: integer
            progress_percentage:
              type: number
            team:
              $ref: '#/components/schemas/TeamMemberList'
    ProjectList:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        total:
          type: integer
    CreateProjectRequest:
      type: object
      required: [name, building_type, leed_version, rating_system, target_level, location]
      properties:
        name:
          type: string
        code:
          type: string
        building_type:
          type: string
        leed_version:
          type: string
        rating_system:
          type: string
        target_level:
          type: string
        location:
          $ref: '#/components/schemas/Location'
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        target_level:
          type: string
    Location:
      type: object
      properties:
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        country:
          type: string
        lat:
          type: number
        lng:
          type: number
    TeamMember:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatar_url:
          type: string
        leed_ap_number:
          type: string
    TeamMemberList:
      type: object
      properties:
        members:
          type: array
          items:
            $ref: '#/components/schemas/TeamMember'
...